app/requirements.txt
black==19.10b0
asgiref==3.2.7
Django==3.0.2
dj-database-url==0.5.0
djangorestframework==3.11.0
flake8===3.7.9
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.1
pytest-cov==2.8.1
pytest-django==3.8.0
pytest==5.3.1
gunicorn==20.0.4
isort==4.3.21
whitenoise==5.0.1